Adding 'tracks'.  This is (or will be) a list of known streetcar tracks, intended to be independent of routes (both our fudgeroutes and NextBus's routes).  The goal of this code is to help us to interpolate vehicle locations better - namely to eliminate the cases in which we currently 'cut corners'  i.e. show a streetcar going through the middle of a block, or on a street that has no streetcar tracks.
We already avoid cutting corners when the two vehicle locations that we are interpolating between are both on-route (see db.py - interp_by_time() - the try_for_mofr_based_loc_interp_ argument).  Unfortunately detours and other oddball situations aren't covered by that.  Hence this 'tracks' code is warranted.

In the future, for buses, another list of all streets (or at least - streets that buses tend to travel on) may be warranted.

 We can tell which vehicles are streetcars because their vehicle IDs start with '4'.

 This is an incomplete list of tracks right now.  It is not used anywhere yet.