Because vehicles advertising an incorrect direction in their dirTag (i.e. containing a _0_ when it should have been _1_, or vice versa) are all too common, I now all but ignore dirTags and instead determine the directions, myself for both traffic and locations.  This warranted a lot of rearranging but it boils down to examining the mofr change between one VehicleInfo and the next to determine the direction.  If the vehicle is off-route (i.e mofr == -1) then I resort to the new 'widemofr' which is mofr with a larger tolerance (currently 2 kilometres) so it will help us determine the direction even if the vehicle is on a detour.
I don't know why vehicles advertise an incorrect dirTag.  I suspect TTC operator error - i.e. this is the NextBus ramification of the common sight of an eastbound streetcar having a westbound (english-language) direction advertised on it's sign out front.  Something that I think reinforces this is that these wrong-direction vehicles seem to happen more often late at night - when, I figure, the vehicles are more empty and transit riders are less likely to notify the driver that their sign is incorrect.

 Headings are also affected by these wrong directions.  They'll be 180 degrees off what one would expect them to be.      I haven't dealt with that yet.