Used to pass a list of list of VIs around.  Was the result of muddled thinking.  Also resulted in traffic reports having gaps between one stretch (i.e. list of VIs) and another.  Simplifying this now.
This idea of a list of list of VIs is only used now when filtering them by direction (db.py line 136) because that's where I make that judgement call that I only want a 'stretch' of VIs of that dir if it's more than 5 readings or 300 metres long.  I do this to filter out 1) fussy readings on a stopped vehicle (where mofr might go 5000, 5001, 5000, 5001, etc.) and 2) parts of detours that look like changes in direction, but that's only because widemofr may (eg. when going eastbound) decrease for a while if the street that is initially being detoured onto is a more-than-90-degree turn from the street being detoured off of.  The latter could happen if for example a southbound Ossington bus detoured westward via Dundas.  widemofr would be increasing on the way down Ossington then decrease on dundas.  This is all so hackish it's barely worth explaining.