I keep seeing faulty headings reported by NextBus.  (This only affects us now for off-route buses.  If a vehicle is on-route or a streetcar, we're already making up our own heading.)
So here I am all but eliminating our use of the headings that NextBus reports.  Now off-route buses' headings are inferred straightforwardly from changes in location, taking the tack that when a vehicle moves from location T to location T+1, it has the heading of the line T->T+1 when it is at T+1.  (It's heading at T is considered to be that of the line T-1->T.)