Optimizing yards (i.e. the removal of vehicles in yards) by adding a bounding box which is checked before the full polygon is.
This sped up a certain test case from 16.5 seconds to 15 seconds.  The test case was path finding, and involved 6 fudgeroutes - which is to say, 6 traffic reports - which is to say, a lot of calls to the function yards.in_a_yard_latlng().  To be exact: 5136 calls.

Before: 0.00118 seconds / call.  After: 0.00043 seconds / call.