Used to be we would prevent all updates of date (traffic and vehicles) from server while paused.  Now with our multiple route selection methods, this doesn't make as much sense.  It was broken.
So now I've eased up on that restriction.  Updates from server are allowed, and I take measures to ensure that g_cur_minute_idx is adjusted to represent the same time it did before the update of vehicle locations - so if we're paused when a locations update arrives, the current minute will not change even though the total time span (eg. 12:07 - 12:37) might change.  Exceptional case is if g_cur_minute_idx was at either extreme (12:07 or 12:37) and that minute disappears after the update.