Fixing vehicle interpolating in the case of having a 'lo' sample but not a high one.  This will happen very often - eg. in a vehicle locations listing for 12:00 - 12:30, most of the 12:30 VIs will be based on a lo sample only, because there is no more recent information.  (Or if we're looking back in time, we're pretending that that there is no more recent information.)
The previous way of doing this wasn't thought through very well, and it looked awful in the case of a vehicle slowing to a stop.  So lolo_vi and lo_vi would have the same latlng, hence the heading between lolo_vi and lo_vi would be more or less random.  (Actually it seemed to work out to 270 a lot for some reason.  Still - not what we want.)

This new way of doing it it though out much better, and uses mofrs and tracks sensibly.  I'm surprised I tolerated the old way for so long.