Adding command-line argument to purgedb.py - 'number of days'.
Getting rid of the full vacuum.  As far as I know the only benefit of that is to return disk space to the operating system,
and now we're off theorem.ca (where my space was tight) that's not necessary.

Also adding an 'analyze' to the vacuum in the hope that it will speed up all the queries.  I have no evidence for this yet.