Re-introducing the concept of 2 memcache instances - 'dev' and 'prod'.
These are both managed from mc.py (called as a main) - eg. "mc.py dev start" on the command line.

Also this commit does away with the .pid files that we used to create on memcache server start,
in order to know what process to kill when we want stop.   Now we find out which process is listening on
the dev or prod port in question, and kill it.