Addendum to previous commit comment:
I don't think it was necessary to switch from lsof to netstat.  It seems that if one runs lsof as root, those processes will show up.  So the last commit was unnecessary.  But this one works just as well - as long as netstat has been setuid'ed.  (lsof would have had to have been too.)