Optimizing the traffic and locations in the typical use case where the client calls both at the same time.  Doing this by using the same db call for both, and memcaching it.  They used the same db function before (db.get_vid_to_vis()) but with slightly different arguments.  It doesn't make a difference because the arguments in question don't make much sense any more.  (I think they made more sense the way the code was many months ago.)  Now all the for_traffic_ argument does is go further back in time ('overshots') if it's false.  So the funny thing is, that argument existed in an attempt at an optimization once.  Now it's better to memcache it.
This speeds up what I'm now calling a 'full cache prime' (that is - getting traffic and locations for all routes and both directions) from 14 seconds to 11 seconds on my home machine.  (Timestamp arg for test was 2013-03-19 20:00).

Also - to really make use of this now-memcached db call sharing - I've got to make the rounding of the time arg the same for traffic and locations.  Right now that means increasing locations' to 60 seconds.  This is unfortunate because I think that users will like to see their vehicles' locations correspond with what they see in real life as closely as possible.  However right now my main goal is to try to make this project public, and on my current web space a full cache prime is often taking over a minute, and that's obviously a show-stopper, so right now optimization is job #1.