Optimizing the main operations (get traffic and get vehicle locations) a lot by reducing the number of additional DB selects ('overshots') that we need to do, by getting a lot more data than we need on the first select.
On the Cirrus Hosting 800Mhz virtual machine, this speeds up a certain full
report get test at from about 45 seconds to 10 seconds, bringing this into the
realm of production use, which is a major milestone.

This optimization had some strange results on locations which I have fixed by
hacking things so that locations are even better than before.

I still barely know what I'm doing but I know that one mistake that I was
making was that if for example a vehicle is going in dir 0 then turns around at
eg. 12:00 and goes in dir 1 (which of course happens all the time) when getting
locations for eg. dir 0 I was trying to interpolate times after 12:00 and of
course, since we had only dir 0 raw locations to work with at that point, this
vehicle turning around looked the same as if it disappeared from the feed.  I
try to account for those scenarios so I was returning locations for it, showing
it at the turn-around lat/lng for 12:01, 12:02, and 12:03 or so.  This was
foolish because in reality, the vehicle was often well on it's way in dir 1 by
12:01.  Now I don't do that by a combination of getting all vis for both
directions, THEN interpolating with all that data at the interpolation code's
disposal, and for this situation and if it is the case, then don't interpolate
there.  (db.py line 578.)

Location results are a little different than before in some cases that don't
matter much, such as now some vehicles being considered dir 0 a.o.t. dir 1 (via
fix_dirtags()) when they're sitting still, about to turn around at the end of a
route.

Traffic seems to be unchanged.