Fixing a problem which I think has only existed for a day or so, but I could be wrong.
Now, when interpolating locations, we are making sure that all times that were
in the time_to_vis_ arg are in the dict that we return, even if we have no vis
left for some of those times.  This is especially important because currently
traffic.php, when showing multiple routes, shows locations only for the
intersection of the times that it has for those routes.  So if we return from
here only a couple of minutes for one of the routes (because for example that
route is going out of service for the day) then that will result in most of the
times for all of the other routes not being shown.  That would be unfortunate.