Fixing major performance problem, which prompted me to take the site down for a couple of weeks. When getting overshots, now selecting on the 'time_retrieved' column instead of 'time'.  Should have done this all along, because there is only an index on 'time_retrieved'.
The select on 'time' was taking several minutes in some cases.

I think that I didn't notice this problem before because the database on this machine used to be small, so scanning
the whole table would only take a few seconds.  Of course it grew, and over the weeks I more and more performance problems.
They were sporadic because that 'overshots' code is not called very frequently, because the initial 'select' (which
has used 'time_retrieved' all along) usually gets enough data.