Introducing the option to use cached pickle files of RouteInfo objects.  This started off as a developer feature.  Each RouteInfo object takes several seconds to build now that their simplified (i.e. different datazoom) versions have to be calculated.  I am introducing these pickle files to make certain kinds of coding workflows faster for me.  This is useful for the kind of scenario where I want to clear eg. traffic or locations results out of memcache but I don't need to clear routeinfo because I haven't changed it.  I can't see how to clear only certain keys out of memcache, so before this commit, after restarting memcache, I would have to wait too long for the routeinfo to rebuild.  If I then loaded traffic.php, I would have to wait for the routeinfos of all routes to rebuild, because of what that page does via php / python. Loading (actually, serving) traffic.php was taking about 55 seconds in this case.
These pickle files will benefit production too, in that if there is ever a
production memcache restart, the first person to load the page after that won't
have to wait for those 55 seconds.  (Not sure how often memcache restarts.
Only if it crashes and my cron job catches that, I think.)  The
reports-generating process will also benefit from this, as it is restarted
sometimes.  (Currently I restart the reports process daily even if it's fine.)