Introducing a Toronto street map (courtesy of Open Data Toronto) and some path-finding code (courtesy of Dijkstra's algorithm).  This is intended for interpolating the locations of vehicles in a better way than what we're doing now.  This code more-or-less works but it isn't used anywhere yet.
I have tested the new snapgraph a little on our streetcar tracks and on the
Toronto map.  For the tracks, I have needed to modify them a little, to a) get
rid of some points that formed loops that the code couldn't deal with, and b)
remove some duplicate sections of track that the code dealt with fine, but
looked ridiculous to me while debugging.

The Toronto map data has, for the sections of it that I have tested so far, no
such loops.  The building of a snapgraph is slow right now.  I think it might
take hours or days to build one for the complete Toronto street map.