Optimizing traffic reports.  get_bounding_mofr_vis() in particular.
This also changes results a little, in a desirable way, I think, in the case
where there is a raw vi that falls on a mofrstep boundary.  For example (if
mofrstep is 100) and we are looking for bounding mofrs for mofr=200 and we have
a raw vis at mofrs 150, 200, and 250 - before this commit we would choose the
150 and 250.  Now we choose 150 and 200.

The behaviour in the last mofr chunk or so at the end of each route is not so
great, neither before nor after this commit.  If a vehicle is sitting in the
station, at mofr 0, for readings, then appears at say mofr 150 (say mofrstep is
100), then we will (I think) consider the bounding vis for mofr 100 to be the
/first/ mofr=0 reading, and the mofr=150 reading.  It would probably be better
if we used the /last/ mofr=0 reading, because our users our probably more
interested in the travel time between mofr 0 and 100 after the vehicle has
left, then the travel time including the waiting time.  This isn't a big deal
now, because for now we're only using these traffic calculations as a visual
indicator and we're probably making more mistakes than this at the ends of the
routes.  But in the future, if we start doing any math based on this traffic
calculation at these end-of-route mofr chunks, then this should be dealt with
somehow.