Optimizing creation of streets snapgraph.   Making one for all of Toronto's streets now takes about 6 minutes on the Cirrus VPS Plan IV.  I don't know how long it took before this commit, because I never did it all.  Somewhere around 10 times as long, I think.
The idea with the change in snapgraph.py (to only yield an addr pair if the
second polylineidx is greater than the first) is because by default this code
yields almost every pair twice - once in one order (addr1, addr2) and once in
the opposite order (addr2, addr1).  So this change makes it yield just one of
those.

The reason it's /almost/ twice is because of what happens when we're yielding
point/point pairs and we're at the end of a polyline.  We yield a little too
much here because the structure we're using to determine which points are in
our gridsquares of interest is gridsquare_to_linesegaddrs - which, true to the
name, identifies linesegs, not points.   I don't feel like explaning this any
more right now.  What can I say.  I think it works.