Working around the secsSinceReport bug.  This is an apparent bug in the XML data from NextBus where secsSinceReport is occasionally a lot high than usual. (80 - 100 instead of the typical 5-20.) This wouldn't be so bad except the implied time (implied through a combination of lastTime and secsSinceReport i.e. lastTime - secsSinceReport*1000) would be farther in the past than the implied time of the previous XML data that we got.  And even THAT wouldn't be so bad, except that the lat/lon in these cases was clearly that of a time /after/ the previous XML data.  That is - it would typically be a lat/lon a few hundred meters down the road, and if we trusted secsSinceReport and interpreted these implied times literally, then (say for a vehicle travelling eastbound i.e. in an increasing mofr direction) this would show up as something like the vehicle being at at mofr=1000 at T-80 seconds, then mofr=800 at T-65 seconds ago, then mofr=1200 at T+55 seconds.  Doubling back like this is physically impossible, especially for streetcars.
So it's as though NextBus was deciding to add about 80 seconds to secsSinceReport every
now and then.  I noticed that this seemed to happen to all vehicles in an XML
document at the same time.

The most noticeable affect for us was that often our 'bad GPS' filtering would
see these jumps in location as implausible, and discard all readings after this
jump.  So vehicles would disappear from the locations report.  (And cease to
contribute to the traffic report.)

This workaround seems to work.  It's not ideal though, because by setting
secsSinceReport to 0, now we're probably creating an implied time that is too
high, and thus showing the vehicles as not as far down the road as they really
are, for a given time.  This could still aggravate the bad GPS filtering, but
on the other side.  Anyway, it's a start.