Tweaking the workaround for what I allege is a secsSinceReport bug in NextBus.
Before this commit I modified secs_since_report only if it clearly caused some
vehicle locations to have times that were out of order.  (i.e w/ two vis, the
second one with the later time_retrieved, having a /combined/ (i.e.
time_retrieved minus secs_since_report) time of earlier than the first vi.)

But that was not enough, because if secsSinceReport from NextBus was freakishly
high but not enough to cause vi2 to have a combined time of less than vi1, then
our workaround code would do nothing.  But then the combined time difference between
vi1 and vi2 would sometimes be say 5 seconds, and the vehicle moved 300 metres,
so it appeared to be travelling 216 kmph.  Then our 'bad GPS group' code would throw out
a lot of these vis.

So I decided that until (or unless) this gets fixed by NextBus, we will ignore
secsSinceReport - or rather, set it to 12, which seems to be the average value
for it.  Vehicle times will be a little off (less than a minute, presumably),
but at least they'll all be there.