Now writing the 'graph_locs' and 'graph_version' columns to the database.  So this introduces the 'graph_locs' member of VehicleInfo.
The value stored in 'graph_locs' is the result of a call to
SnapGraph.multisnap() (on the appropriate snapgraph - tracks or streets), and
we are storing it in the database, at 'poll locations' time, for performance.

Also starting with this commit, these new values are taken advantage at report
generation time, for performance.

I haven't tested either of these parts yet, because I don't know how to test
these things before committing them.