Fixing detouring 503 and 508 vehicles.  Before this commit, each 503 or 508 vi would be assigned to queen or king based on which fudgeroute first yielded a mofr != -1 for the vi's latlng, or if neither, the first in the list would be chosen (which happened to be queen).  For a 'detouring' 503 vehicle that is, say, going west along Wellington around Bay, before this commit the fudgeroute would be set to queen.  Now it's king.
This resulted in some vehicles appearing to change between fudgeroutes over
time without any apparent reason, and so would disappear from location reports
without any clear reason.

('detouring' above is in quotes because it's not really detouring according to
the TTC.  They all go along Wellington in this area.  It's only detouring in
the sense that it's not "on" (which currently means: within 50 meters of) our
fudgeroute king.)