Optimizing LatLng.__eq__().  Speeds up locations reports by about 5%.
Right now I think that LatLng.__eq__() is called a lot because
db.get_grade_stretch_info_impl() is using lru_cache w/ a list of VehicleInfos
as an argument, so there's a lot of vis (and hence latlngs) in the lru_cache
dict.  db.get_grade_stretch_info_impl() probably won't be using lru_cache for
much longer because I hope to make a patch-cache version soon.  But for now at
least, this speeds things up a bit.