1) Failed experiment in optimizing SnapGraph.find_multipath() with a patch cache.
The patch cache version ran at almost exactly the same speed as the regular
one.

Committing this code, disabled, in case I want to go back to it later.

Maybe this is a useful refactoring of this function anyway.

Also, some minor changes occurred to some code (but not results, I think), such
as PosAddr became immutable.


2) calc-reports.py got spruced up with the --prof option, and using python's -O
(optimize) flag or not, depending on the context.


3) "cputime" got the --optlevel option.