Optimizing db.interp() with a patch-cache.  This reduces the CPU time that locations reports take to generate by about 35%.
I think that commit comments before this one mentioning optimization
percentages are mostly over-reporting, because I didn't think to generate all
datazooms when testing, and a lot of those commits optimized only (or
primarily) functions that are called only once and re-used by all datazooms.

But for this commit and this figure of 35%, I tested all datazooms.