Simplified route points (for each datazoom) are now calculated from the raw route points with the RDP algorithm, rather than the old home-made (and not very good) algorithm.
This results in a lot fewer points for the higher datazooms.

Doing this not for its own sake, but because I want to do something else with
route points soon, that the old home-made algorithm didn't allow.  In
particular: for split-by-dir routes, cheating to make a stretch of road longer
by introducing very small (< 1 meter) zig-zags.  The old algorithm didn't seem
to allow raw points that were less than one meter apart, and it didn't seem to
simplify away zig-zags that were much larger (10-50 meters?)