Paying attention to HTTP caching for the first time.  The max-age of everything is now 1 day, but that is disabled for AJAX calls (or rather - thwarted due to the timestamp appended to the retrieved URL, thanks to this new ajaxSetup call).
Also, in conjunction with this, the main page (traffic.php) automatically
reloads every 24 hours now.

Doing this because of some error logs that were seen on the server.  It looked
like some users were browsing copies of traffic.php that were several days old
and didn't match the server-side code (called via ajax / callpy.wsgi) any more.
So the point of this commit to put a time limit on how long after a site update
that kind of thing happens.  That particular client page / server code version
mismatch in the logs didn't degrade the user's experience, but the next one
might.

I'm going on the assumption that browsers might cache forever if there are no
cache headers (as there weren't, before this commit.)