Fixing a bug where the user would not be able to get a list of routes to/from certain locations.  They would see the error icon, and an error would appear in the apache logs.  I forget what kind of error, but this assertion at line 388 of system.py would fail: 	assert path[0] == orig_boundingvertidx and path[-1] == dest_boundingvertidx ... and even if assertions were disabled, the data retrieved would be a mess from that point onward.
This bug happened because parts of the system graph were broken.  The memcache
restart that this commit introduces while  building the system graph is
supposed to fix this.  I never caught this bug red-handed.  Hopefully
SystemSnapGraph.check_pcp() will catch these errors if I'm wrong about the
memcache cause, and these kinds of errors happen again.

Some other things are changing in the building of graphs during this commit.
Each graph built, and all of the vertexes and polylines within it, are supposed
to be consistent now for multiple builds from the same source data.  This goes
for all graph, not just the system graph.  This is intended to make the
debugging of graph building (i.e. future problems like the one above) easier.
Attempts were made to do this before this commit, but they weren't very good.